requirements.txt
streamlit
scikit-learn
pandas
numpy
matplotlib
newspaper3k
Sastrawi
